move track design related functions to new source
