+++ openrct2.vcxproj
@@ -181,6 +181,7 @@
     <ClCompile Include="src\ride\thrill\twist.c" />
     <ClCompile Include="src\ride\track.c" />
     <ClCompile Include="src\ride\track_data.c" />
+    <ClCompile Include="src\ride\track_design.c" />
     <ClCompile Include="src\ride\track_list.c" />
     <ClCompile Include="src\ride\track_paint.c" />
     <ClCompile Include="src\ride\transport\chairlift.c" />
@@ -375,6 +376,7 @@
     <ClInclude Include="src\ride\station.h" />
     <ClInclude Include="src\ride\track.h" />
     <ClInclude Include="src\ride\track_data.h" />
+    <ClInclude Include="src\ride\track_design.h" />
     <ClInclude Include="src\ride\track_paint.h" />
     <ClInclude Include="src\ride\vehicle.h" />
     <ClInclude Include="src\ride\vehicle_paint.h" />
