+++ src/object_list.c
@@ -16,13 +16,14 @@
 
 #include "addresses.h"
 #include "config.h"
+#include "game.h"
 #include "localisation/localisation.h"
 #include "object.h"
 #include "platform/platform.h"
-#include "ride/track.h"
-#include "util/sawyercoding.h"
-#include "game.h"
 #include "rct1.h"
+#include "ride/track.h"
+#include "ride/track_design.h"
+#include "util/sawyercoding.h"
 #include "world/entrance.h"
 #include "world/footpath.h"
 #include "world/scenery.h"
