+++ src/windows/track_list.c
@@ -17,15 +17,16 @@
 #include "../addresses.h"
 #include "../audio/audio.h"
 #include "../editor.h"
-#include "../localisation/localisation.h"
+#include "../interface/themes.h"
 #include "../interface/widget.h"
 #include "../interface/window.h"
+#include "../localisation/localisation.h"
+#include "../rct1.h"
 #include "../ride/ride.h"
 #include "../ride/track.h"
+#include "../ride/track_design.h"
 #include "../sprites.h"
 #include "error.h"
-#include "../interface/themes.h"
-#include "../rct1.h"
 
 enum {
 	WIDX_BACKGROUND,
